# src/openApi/swagger.json
{
    "openapi":"3.0.0",
    "paths":{
       "/redocly":{
          "post":{
             "operationId":"RedoclyController_create",
             "parameters":[
                
             ],
             "requestBody":{
                "required":true,
                "content":{
                   "application/json":{
                      "schema":{
                         "$ref":"#/components/schemas/CreateRedoclyDto"
                      }
                   }
                }
             },
             "responses":{
                "201":{
                   "description":""
                }
             }
          },
          "get":{
             "operationId":"RedoclyController_findAll",
             "parameters":[
                
             ],
             "responses":{
                "200":{
                   "description":""
                }
             }
          }
       },
       "/redocly/{id}":{
          "get":{
             "operationId":"RedoclyController_findOne",
             "parameters":[
                {
                   "name":"id",
                   "required":true,
                   "in":"path",
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":""
                }
             }
          },
          "patch":{
             "operationId":"RedoclyController_update",
             "parameters":[
                {
                   "name":"id",
                   "required":true,
                   "in":"path",
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "requestBody":{
                "required":true,
                "content":{
                   "application/json":{
                      "schema":{
                         "$ref":"#/components/schemas/UpdateRedoclyDto"
                      }
                   }
                }
             },
             "responses":{
                "200":{
                   "description":""
                }
             }
          },
          "delete":{
             "operationId":"RedoclyController_remove",
             "parameters":[
                {
                   "name":"id",
                   "required":true,
                   "in":"path",
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":""
                }
             }
          }
       }
    },
    "info":{
       "title":"Redocly  example",
       "description":"The Redocly  API description",
       "version":"1.0",
       "contact":{
          
       }
    },
    "tags":[
       {
          "name":"Redocly ",
          "description":""
       }
    ],
    "servers":[
       
    ],
    "components":{
       "schemas":{
          "CreateRedoclyDto":{
             "type":"object",
             "properties":{
                
             }
          },
          "UpdateRedoclyDto":{
             "type":"object",
             "properties":{
                
             }
          }
       }
    }
 }